LPD-79027 Move functions that update state to common folder and rename some of them
