+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/utils/state/addRepeatableGroup.ts
@@ -3,15 +3,19 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {RepeatableGroup, Structure, StructureChild} from '../types/Structure';
-import {Uuid} from '../types/Uuid';
-import getRandomId from './getRandomId';
-import getRandomName from './getRandomName';
+import {
+	RepeatableGroup,
+	Structure,
+	StructureChild,
+} from '../../types/Structure';
+import {Uuid} from '../../types/Uuid';
+import getRandomId from '../getRandomId';
+import getRandomName from '../getRandomName';
 import sortChildren from './sortChildren';
 
 const DEFAULT_GROUP_LABEL = Liferay.Language.get('repeatable-group');
 
-export default function addGroup({
+export default function addRepeatableGroup({
 	groupChildren,
 	groupParent,
 	groupUuid,
@@ -39,7 +43,7 @@ export default function addGroup({
 		if (child.type === 'repeatable-group') {
 			const group: RepeatableGroup = {
 				...child,
-				children: addGroup({
+				children: addRepeatableGroup({
 					groupChildren,
 					groupParent,
 					groupUuid,
