+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/utils/state/sortChildren.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {Structure, StructureChild} from '../types/Structure';
+import {Structure, StructureChild} from '../../types/Structure';
 
 export default function sortChildren(
 	children: Structure['children']
