+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/utils/state/updateHistory.ts
@@ -3,12 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {State} from '../contexts/StateContext';
-import {Structure} from '../types/Structure';
-import {Uuid} from '../types/Uuid';
-import findChild from './findChild';
+import {State} from '../../contexts/StateContext';
+import {Structure} from '../../types/Structure';
+import {Uuid} from '../../types/Uuid';
+import findChild from '../findChild';
 
-export default function getHistory({
+export default function updateHistory({
 	deletedChildrenUuids,
 	initialHistory,
 	publishedChildren,
